+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state.dart
@@ -130,7 +130,7 @@ abstract class MPTH2FileEditState {
 
   void setCursor() {}
 
-  void setStatusBarMessage() {}
+  void updateStatusBarMessage() {}
 
   void onPrimaryButtonPointerDown(PointerDownEvent event) {}