+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state_select_empty_selection.dart
@@ -23,7 +23,7 @@ class MPTH2FileEditStateSelectEmptySelection extends MPTH2FileEditState
       selectionController.clearSelectedEndControlPoints();
       selectionController.clearSelectedLineSegments();
     }
-    setStatusBarMessage();
+    updateStatusBarMessage();
   }
 
   @override
@@ -35,7 +35,7 @@ class MPTH2FileEditStateSelectEmptySelection extends MPTH2FileEditState
   }
 
   @override
-  void setStatusBarMessage() {
+  void updateStatusBarMessage() {
     th2FileEditController.setStatusBarMessage(
       mpLocator.appLocalizations.th2FileEditPageEmptySelectionStatusBarMessage,
     );